fundamental changes to the build system: let make and gcc figure out the dependencies
With this changes is should not be necessary anymore to declare any source or header dependencies.
Of course you still need to declare the .o or .a dependencies for the libraries or executables.
But everything else should work automatically from there on.

There are two rough edges, because the .P files are collected in dep/ and all of them are included:
- Avoid two .cc files with the same name.
- When you delete a .cc file, there will be a stale .P file in dep/ that might or might not cause problems.
  In case this happens, just delete the offending .P file or do a make clean.
  NEVER delete the whole dep/ directory without doing a make clean.