Created new configuration file management called MConfig.
It is a part of the Tools namespace now. The main functions are templated so most of the code is and has to be in the header file.
MConfig scans the config file it was initialised with for imports of subsequent config files.
The (default) keyword to import one config file into another is 'import' followed by the path of the file to be included in (double) quotation marks.
The import is done recursively.
Accordingly the cut files have been modified so that all import 'CommonCuts.txt' now.

With the change to the new config class, it isn't necessary to use ROOT file types any more (e.g. you can use std::string instead of TString now when extracting information from the config class).

Also adapted submit_music.py for Condor to transfer the whole ConfigFiles directory, so files imported from configuration files can be found.