Fix floating point computation
In certain rare cases comparing a float temporarily stored in the RAM with one kept in the FPU leads to unexpected results.
The -ffloat-store option causes gcc to channels all floats through RAM or registers, so all of them are handled indentically.
This might lead to some loss in precision, but gains in intuitive behavior.