New JES uncertainty treatment
At this moment (CMSSW 3_8_7) the pt and eta dependent corrections of jet pt can be retrieved from textfiles.
The path of these textfiles can be specified in the config file and are in general different for data and MC.

The varyJES and varyJESMET methods in the EventSelector have been adapted to handle the jet pt and eta dependent corrections.