Rebuild the command line of music.
There were several relicts in the music options that have been removed now.

One strict rule is now and should always be:
All analysis based options and configurations belong into the configuration
files, not into the command line of music!

Also the number of options for music should stay short, because they must be
propagated to radio.py.

In the course of this, there have also been changes to radio and the Condor and
GRID submission scripts.

radio.py: - Acts now as a control script for Condor and GRID job submission. It
            calls condor_radio.py and grid_radio.py, respectively, depending on if you
            want the classification to be done on the GRID or on Condor (or
            both). Therefore the submit scripts have been rewritten in Python.
            The actual structure of GRID and Condore jobs has not changed
            (i.e. for Condor jobs you "just" have to wait, for GRID jobs you
            must use jukebox as before).
          - A new split jobs option allows to run the classification of one
            skimmed process on more than one core. You have to specify the
            number of pxlio files per job.
            If you ran jobs in split-mode you must use ECMerger2 to merge them.