MISPrinter: Print also Data/MC ratio.
Had to restructure a big part of the code.
Many calls are in (member) functions now.

The new plots should work with every ROOT version now
(Segmentation Violation in some cases with the old plots).

The crash when calling the 'Print' function of TCanvas with num_dice = 0
should not appear anymore.