EventCleaning: Updates for muon collection cleaning.
It is not perfectly clear, if we really want to clean muons against muons. In
general, there can be ambiguities in the reconstruction (ghost), but e.g. for
high-pt Z candidates you expect two muons to be very close in Delta R.
TODO: We need a study/reference here.

Some more info:
https://twiki.cern.ch/twiki/bin/view/CMSPublic/SWGuideMuonAnalysisMuonAssociators?rev=33#Muon_Classification_By_Hits

If the duplicate removal is active (config file), check probability for muon
tracks rather than the chi2/ndf.