WIP: Workaround for events with no unprescaled single muon or electron triggers.
There were some events during several periods of Run A where there are no
unprescaled electron or muon triggers. Perhaps a special prescale column was set
during these runs, but we do not have this information in our skimmed datasets
yet. Until we adjust the skimmer and add the column information to the skimmed
datasets we can use this workaround in the classification.
Each RECO  event is explicitly checked for unprescaled triggers beginning with
'HLT_HLT_Ele' or 'HLT_HLT_IsoMu'. If none are present a warning is isued and the
event is skipped.