WIP: music: Workaround for crash caused by misformatted pxlio files.
From time to time (rarely) pxlio files on the dCache are formatted in such a
way, that the last event is not closed properly. The event loop in music.cc will
then try to open the following event, which leads to an exception.
By checking the size of the event's UserRecord (should be > 0) these instances
can be caught and skipped.