EventSelector: New cut on number of tracks in the event.
There are events with very odd jets that originate from "too many" tracks in the
tracker. These tracks are unphysical and the resulting jets are not real.
Before CMSSW_5_X_Y there are no working filters to get rid of these events on
skimmer level. So this is a workaround.