Disentangled EventSelector and ParticleMatcher.
ParticleMatcher works standalone now and is not called via EventSelector anymore
(there was no reason to do so).

To make it work, a new class was introduced 'GenRecNameMap'.
In general, it is possible that physics objects are labeled differently in Gen
and Rec events (e.g. 'genMetCalo' and 'pfMETNoPU').
The new class reads the config file and creates an (iterable) mapping between
the 'GenName' and the 'RecName' of the objects.
Thus, the user is responsible for the right choice of object collections to be
used here, especially for matching.

Another consequence is that the old 'ControlPlotFactory' is not reliable
anymore because it was never designed to handle such "complex" behaviour.