Restructured config files!
Most config files have been renamed and a new structure is used now:
To run the classification (the program called 'music'), you need a "complete"
config file (a .cfg file) that is build upon a config fragment (.cff file).

The config fragments (.cff) provide configuration that is the same in all
classification ways (Data, MC, FASTSIM).

The final config files (.cfg) define everything that is specific for the given
classification way (e.g. Trigger configuration for MC).
Only a fully functional config file is meant to be a .cfg file!

At this moment, the structure is the following:

   - Trigger.cff: Defines all trigger groups that can be used within 'music' but
     does not specify the trigger paths to be used.

   - Classification.cff: Defines general classification settings and the full
     event and object selection.
     Imports (uses) 'Trigger.cff'.

   - Data.cff: Specifies everything that is relevant for classification of Data.
     Especially sets the trigger paths to be used in all Data samples (but not
     the trigger configuration).
     Imports 'Classification.cff'.

   - Data_<datastream>.cfg: For each datastream (SingleMu, SingleElectron, etc.)
     there is a .cfg file that defines the final trigger configuration. This can
     be used for classification of Data samples.
     Import 'Data.cff'.

   - MC.cfg: Specifies everything that is relevant for classification of MC.
     Especially sets the trigger paths AND the trigger configuration to be
     used in MC. Imports 'Classification.cff' and can be used for
     classification of MC samples.

   - MC_<special>.cfg: Special MC config files that can define various
     (additional) settings for MC samples. At this moment, these are used to
     define invariant mass cuts to be applied on Generator level OR to define
     filtering schemes to "cross clean" QCD samples (QCD EM, mu enriched samples
     etc.).
     Import 'MC.cfg'.

'mkSubmitFile' should be superfluous now.

Free Bradley Manning!