WIP: EventSelector: Get the RECO photon's eta via supercluster information.
As is the case with electrons, it is recommended to determine the pseudorapidity
of a photon via its supercluster information, not via its 4-vector.

The Skimmer has already been updated, but the samples have not been reskimmed
yet. Update the code (delete the try-catch statement) once all skimmed samples
include the variable 'SCeta' for photons.