WIP: EventSelector: Made JES correction uncertainty type configurable.
The Jet Energy Scale Correction uncertainties are given in a text file, which
contains different types of uncertainties. It is recommended to use the 'Total'
uncertainty for JES variations, but the individual uncertainty sources are also
available.
Which uncertainty source(s) you want to take into account MUST now be set in the
'Classification.cff'. For details see:
https://twiki.cern.ch/twiki/bin/viewauth/CMS/JECUncertaintySources?rev=19#Main_uncertainties_2012_53X