WIP: EventSelector: Added possibility to use High Pt Muon ID.
For muons with pt>200 GeV you can now use the Hight Pt Muon ID by setting the
bool 'Muon.HighPtMuonID.use' in your 'Classification.cff' to True. For muons
with a smaller pt the High Pt Muon ID is NEVER used.
When using the High Pt Muon ID, make sure to also set 'UseCocktail' to true, so
the correct muon pT is used.

Also fixed the skimmer variable for the longitudinal impact parameter 'dz' used
in the Tight Muon ID. The wrong muon track was being used previously.

Furthermore, the cut on the relative error of the global muon track was removed
for the Tight Muon ID. This cut was not from the official ID and has been
incorporated in the new High Pt Muon ID.

New variables were added to the skimmer to get both IDs. Once all samples have
been re-skimmed with these variables remove the try-catch blocks from the code.