Install LHAPDF6 and prepare everything to run with it!
installLHAPDF: Script to perform all needed steps to download and install LHAPDF
locally (includes download and installation of yaml).

In '$MUSIC_BASE/external', the downloaded source files are located. It appears
meaningful to have a copy of them inside the repo for all cases.

The LHAPDF libraries will be installed to '$MUSIC_BASE/local'.