WIP: plotPtilde: Added option to disable the interactive mode.
This is very helpful if you want to execute 'plotPtilde' several times, i.e. in
a script.

Apparently there is a ROOT bug concerning drawing the THStack 'p_tilde_stack' a
second time to re-draw the axis and potential grid lines. As a result the script
will not end properly and your CPU usage will increase to about 100%.
Fortunately this problem does not occur if the TH1F 'p_tilde' is drawn a second
time. (Why ROOT? Why??)