JetResolution: New Class to read jet resolution corrections that will be applied in EventAdaptor.
This class parses a given config file and provides resolution correction factors
depending on a given Rec jet and its matched Gen jet.

In EventAdaptor all jets in each event are "smeared" depending on the matched
jet. The changes in the jet pt are added up and also applied to the MET in the
event.

We follow the recipe from:
https://twiki.cern.ch/twiki/bin/viewauth/CMS/JetResolution?rev=32