EventSelector: Updated 'checkOrder()' function.
Throw an exception instead of exiting the program brutally.

Additionally, when running on MC and a particle is found to be unsorted (by pt),
catch the exception (in main()) and ignore the whole event.
(Seems to be a PXL problem.)
Not caught, when running on data. If this happens on data, you should
investigate!