WIP: aix3adb2music: New script to fill database information into MUSiC configs.
Use this script to get information from the dataset database at:
   'http://cern.ch/aix3adb'
and automatically fill it into classification config files.

The files 'XSections.txt' and 'ErrorContributions.txt' have been combined to a
new file 'Scales.txt'. This will hopefully make the accessing the information
quicker. In 'music' and the 'MISMaster' the defaults are still set to the old
config files, so make sure to use the respective options and specify the new
'Scales.txt' config file.

Furthermore a template can be made for 'Colors.txt', which is meant to be kept
separate, since each user might have several different versions of this file.