WIP: submitClassification: New script to submit class. jobs.
Use this script together with config files created with, e.g.,
'aix3adb2music.py' to submit data and MC classification jobs.

'submitClassification.py' parses the datasets from the given playlist and fills
the submission info into a dict. Since python 2.6.4 (the version that comes with
CMSSW 5_3_14) does not yet contain the possibility to use an orderedDict (python
verion >= 2.7) the samples would not be in the order they are given in the
playlist. This means that if the submission breaks down for what ever reason,
you cannot easily findout which samples have already been submitted.
As a workaround an official method to use the orderedDict in python 2.6 has been
added to our repository. More infos here:
   http://code.activestate.com/recipes/576693/

Config files for 2012 will be added in a later commit.