EventSelector/EventAdaptor: Bugfix for propagating jet corrections to MET.
To propagate the jet correction correctly to MET it is important to realize that
the components (x,y) of MET and of the sum of the corrections are both signed
numbers. To take all four combinations into account the corrections must be
subtracted from the MET (taking the signs into consideration).

This is a direct fix for commit:
   3e238096f54a7b61094f8dd806dbe4e074b4c54f
and also:
   aefb2f61a184d6c1f67813a1934c6949c48fffa8