GenSelector: New class for cutting on generator variables.
If you want to use generator cuts set 'Generator.use' to true in the
corresponding classification config.

At the moment cuts can be applied to the invariant mass and the transverse
momentum of a resonance particle. Additional variables can now be added fairly
easily in 'GenSelector.cc'.

Added configs for several tail samples:
   * ttbar binned in mass(ttbar)
   * Wjets binned in pt(W)
   * DY binned in pt(Z)
and included them in 'playlist-mc.txt'.

Furthermore these Gen variables are now plotted in the Gen ControlPlots
(GenKinematicsControl). Add any additional variables here too.

Notice that for pt(W) and pt(Z) the cuts do not result in a steep (vertical)
fall, since jets added in the pythia showering will shift the boson's pt to
accommodate the extra jets. In the config files the cut is still applied on the
unbinned sample at the exact value, since only about 5-10% of events will be
shifted above or below the cut value.