(Validator) Made the code runable with the first version of the validator
Included many default values in the Framework and specialAna to work
the analysis with old (2012) Config files. Also included the first
version of the validator with the basic necessary functions (included
as empty functions at the moment). At the moement it runs the analysis
and keeps track of the memory usage of the analysis.