(Validator) Completed the mememory and run time comparison
The memory usage and run time are now completely compared
between the reference and the new measurment. The control
plot is now done with matplotlib, and stored in the comparison
directory. The tolerance which is accepted by the program
can be defined as command line options, the default values
are 10% for the memory usage and 100% for the run time.