(Validator) Include validation make target
To compile the validation analysis just type 'make validation',
then the analsysis from the folder Validator/ is used in the
compilation of the framework. If you just type 'make' the
folder specialAna/ is used. The folder specialAna/ is no longer
tracked in the repository, so that it is able to just include
your own analysis. If the folder does not exist it will be
created by the Makefile and the files from the Validator are
copied in the folder as a first dummy analysis.