(Validator) Included User decision
If there are deviations the user has to decide if they are acceptable,
therefore the summary .pdf file is opened and the user has to have a
look. If the user decision is positive he has to type in the control
sequence from the .pdf file. Otherwise the validation has failed.