Updated the validator to work in batch mode
To not display all created plots of matplotlib the backend was chosen
to be 'AGG'. Therefore the plots are now only shown in the summary file.
Also the default command in z_dummy.sh was updated to use the latest
config file. And also it is checked if there is a log/ folder and it is
created if not.