Adapt music.cc and Selection to new trigger logic in skimmer
The recent version of the skimmer used an adapted logic to store
trigger information. The new logic introduces a new event view
"Trig". This event view contains one user record per fired HLT
trigger. Previous versions saved one userRecird for each possible
 trigger and a bool whether it has fired or not.

This commit adds the new event view to the main event loop
in music.cc and propagates it to the selection functions.

The main change of the logic to check if a Trigger has fired can
be found in Main/TriggerGroup.cc.