New Ele Selection logic implemeted
The new logic introduces the possibility to choose if cuts
should be expliciltly checked or if the pre checked selection
from the skimmer (new) shoukd be used  or if selection
should be based on cuts defined in classification config.
Cut based (CBID) and HEEPID are available.
Use e.g.
Ele.CBID.usebool = 1
Ele.CBID = "IDname"
To use the skimmer preselection saved as user record "IDname" for ID.