Adapt EffectiveArea to new config structure
The EffectiveArea Classes are adjusted to the new config structure.
Effective areas are now part of the corresponding onject fragment.
The input string for the previous config files is replaced by a
string for the object type.
Variable replacements:
+ ObjectName + ".EffArea.eta_edges",
+ ObjectName + ".EffArea.EA_neutral_hadrons",
+ ObjectName + ".EffArea.abs_eta"
in code corrspond e.g. to Jet.EffArea.abs_eta in jet.cff