Introduce OldNameMapper for changed UserRecord keys
This commit indtroduces the OldNameMapper class.
It can be used to keep the PxlAnalyzer code compatible for
pxlio files created with different versions of the PxlSkimmer.

If a user record name key has changed for a variable while
THE DEFINITION OF THE VARIABLE REMAINS SAME one can replace
the user record while destroying the ability to run on pxlio
files which saved the data with a different keys.

Two kinds of replacements exist:
* General replacements which are valid for all kinds of objects
* Object specific relacements which are only valid for a object.
  This allows to have two different replacments for a key
  dependent on the object type.

Replacements are added in the constructor of the oldNameMapper
Class in chronological order w.r.t. the changes in the PxlSkimmer.
This allows to update the cache for replacements only once with
all needed changes if a requested replacements suggests that the
pxlio file uses a older pxlio skimmer version as detected before.

Here is a example:
double const variable = part->getUserRecord(
        oldNameMap.getUserRecordName( part, "Gamma", "sigma_iEta_iEta" )
);
This will consider the name sigma_iEta_iEta or will try to use one
of the existing replacements.

The constructor of the oldNameMapper contains the line:
addReplacement( "Gamma", "sigma_iEta_iEta", "iEta_iEta" );
and allows to run on samples where the user record was named iEta_iEta.