Added option to separate bJets from Jets
This commit introduces the option Jet.BJets.separate.
The JetType writer is adjusted to change the PxlParticle name
to b. The list of particle abreviation uses bJets as a option now.
This leads to two seperate entries in particle / use / count maps
created by the corresponding functions in the EventSelector.
The position for jet type writing in music.cc had to be changed to allow
the TypeWriting after the selection, because bJets should still be
treated as a common jet in the selection.