Systematics: differentiate between constant and differential systs
Added option to add constant systematics with a fixed shift value.
Here is the idea: All systematics are added as SystematicsInfo objects
in the m_activeSystematics variable in the Systematics class. The
instance of this class is available from the AnalysisCompaoser
to all individual Analysis. In your Analysis you may loop over
all systematics and do one of the following:
* Calculate the shifted value for the variable you work on from the
  shifted views from the eventView pointers stored for a DIFFERENTIAL
  systematic.
* Fill the value for the studied variable calculated from the
  standard view (e.g. Rec) event view in Fill it with a constant
  shift, which is saved in the Systematics Info object for CONSTANT
  systematics.