Systematics: fix potential seg fault in createFullViews
The createFullViews functions deleted particles during a loop
and tried to continue to work on the object. Most time this
bug seems to be catched by pxl. This is why it was not noticed
during testing, but appeared  in full runs.
This commit moves the deletion after the inner loop-